Move `Kafka` from `zio.kafka.testkit.embedded` to `zio.kafka.testkit`
