+++ zio-kafka-bench/src/main/scala/zio/kafka/bench/ConsumerBenchmark.scala
@@ -3,12 +3,12 @@ package zio.kafka.bench
 import io.github.embeddedkafka.EmbeddedKafka
 import org.apache.kafka.clients.consumer.ConsumerConfig
 import org.openjdk.jmh.annotations._
-import zio.kafka.testkit.KafkaTestUtils.{ consumer, produceMany, producer }
 import zio.kafka.bench.ZioBenchmark.randomThing
 import zio.kafka.consumer.{ Consumer, Offset, OffsetBatch, Subscription }
 import zio.kafka.producer.Producer
 import zio.kafka.serde.Serde
-import zio.kafka.testkit.embedded.Kafka
+import zio.kafka.testkit.Kafka
+import zio.kafka.testkit.KafkaTestUtils.{ consumer, produceMany, producer }
 import zio.stream.ZSink
 import zio.{ durationInt, Ref, Schedule, ZIO, ZLayer }
