+++ zio-kafka-test/src/test/scala/zio/kafka/AdminSaslSpec.scala
@@ -3,7 +3,6 @@ package zio.kafka
 import zio._
 import zio.kafka.admin.acl._
 import zio.kafka.admin.resource.{ PatternType, ResourcePattern, ResourcePatternFilter, ResourceType }
-import zio.kafka.testkit.embedded.Kafka
 import zio.kafka.testkit._
 import zio.test.Assertion._
 import zio.test.TestAspect._
