+++ zio-kafka-test/src/test/scala/zio/kafka/consumer/SubscriptionsSpec.scala
@@ -3,12 +3,11 @@ package zio.kafka.consumer
 import io.github.embeddedkafka.EmbeddedKafka
 import org.apache.kafka.clients.consumer.ConsumerConfig
 import zio._
-import zio.kafka.testkit.KafkaRandom
-import zio.kafka.testkit.KafkaTestUtils._
 import zio.kafka.TestLogger.logger
 import zio.kafka.producer.Producer
 import zio.kafka.serde.Serde
-import zio.kafka.testkit.embedded.Kafka
+import zio.kafka.testkit.{ Kafka, KafkaRandom }
+import zio.kafka.testkit.KafkaTestUtils._
 import zio.stream.{ ZSink, ZStream }
 import zio.test.Assertion._
 import zio.test.TestAspect._
