+++ zio-kafka-test/src/test/scala/zio/kafka/utils/SslHelperSpec.scala
@@ -6,8 +6,7 @@ import zio.ZIO
 import zio.kafka.consumer.{ Consumer, Subscription }
 import zio.kafka.producer.Producer
 import zio.kafka.serde.Serde
-import zio.kafka.testkit.embedded.Kafka
-import zio.kafka.testkit.{ KafkaRandom, KafkaTestUtils }
+import zio.kafka.testkit.{ Kafka, KafkaRandom, KafkaTestUtils }
 import zio.test.Assertion._
 import zio.test.TestAspect._
 import zio.test._
